feat(list): List files in archive without decompressing them
basic support

fix #23

+++ src/test/java/nc/opt/uil/j7zip/Tests.java
@@ -10,11 +10,14 @@ import java.io.IOException;
 import java.io.PrintStream;
 import java.nio.file.Files;
 import java.nio.file.Paths;
-import nc.opt.util.J7zip;
+import java.util.stream.Stream;
+
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.opentest4j.AssertionFailedError;
+
+import nc.opt.util.J7zip;
 import picocli.CommandLine;
 
 public class Tests {
@@ -72,6 +75,39 @@ public class Tests {
         assertArrayEquals(Files.readAllBytes(Paths.get("src/test/resources/poem.txt")), Files.readAllBytes(Paths.get("target/poem.txt")));
 
         Paths.get("target/poem.txt").toFile().delete();
+        Paths.get("target/archive.7z").toFile().delete();
+    }
+
+    @Test
+    public void testListArrchiveFiles() throws IOException {
+        // compress
+        int exitCode = new CommandLine(new J7zip()).execute(new String[] { "a", "target/archive.7z", "src/test/resources" });
+        restoreStreams();
+
+        String err = new String(this.err.toByteArray());
+        if (!err.isEmpty()) {
+            System.err.println(err);
+            throw new AssertionFailedError(err);
+        }
+        assertEquals(0, exitCode);
+
+        // list contents of file
+        setUpStreams();
+        exitCode = new CommandLine(new J7zip()).execute(new String[] { "l", "target/archive.7z" });
+        restoreStreams();
+
+        err = new String(this.err.toByteArray());
+        if (!err.isEmpty()) {
+            System.err.println(err);
+            throw new AssertionFailedError(err);
+        }
+        assertEquals(0, exitCode);
+
+        String[] lines = new String(this.out.toByteArray()).split("\n");
+        assertEquals(2, lines.length);
+        assertTrue(Stream.of(lines).anyMatch(f -> f.equals("src/test/resources/poem.txt")));
+        assertTrue(Stream.of(lines).anyMatch(f -> f.equals("src/test/resources/poem-with-password.7z")));
+
         Paths.get("target/archive.7z").toFile().delete();
     }
 
@@ -150,8 +186,7 @@ public class Tests {
     @Test
     public void testDecompressWithoutTarget() throws IOException {
         // decompress
-        int exitCode = new CommandLine(new J7zip())
-            .execute(new String[] { "e", "-p", "poem", "src/test/resources/poem-with-password.7z" });
+        int exitCode = new CommandLine(new J7zip()).execute(new String[] { "e", "-p", "poem", "src/test/resources/poem-with-password.7z" });
         restoreStreams();
 
         String err = new String(this.err.toByteArray());
@@ -163,10 +198,7 @@ public class Tests {
 
         // check (compare decompressed file with initial)
         assertTrue(Paths.get("poem.txt").toFile().exists());
-        assertArrayEquals(
-            Files.readAllBytes(Paths.get("src/test/resources/poem.txt")),
-            Files.readAllBytes(Paths.get("poem.txt"))
-        );
+        assertArrayEquals(Files.readAllBytes(Paths.get("src/test/resources/poem.txt")), Files.readAllBytes(Paths.get("poem.txt")));
 
         Paths.get("poem.txt").toFile().delete();
     }
